Refactor into http module
